fix: adjust genderId datatype in command

+++ src/main/java/io/github/ctorressoftware/academic/enrollment/person/application/port/in/create/CreatePersonCommand.java
@@ -9,6 +9,6 @@ public record CreatePersonCommand(
         String lastName,
         String secondLastName,
         Document document,
-        int genderId,
+        short genderId,
         Email email
 ) {}
